Refactor 05_sumAll solution and add new test cond.
My code, whilst not the same as the solution still
managed to pass all the tests... however I noticed
that my counter variable started from 0 each time
instead of from the smallest integer. This worked
fine for the tests since no test tested for any
minimum values greater than 1 that were a
positive whole integer. I created a new test
condition to test for this starting at 4 as the
min and 8 as the max.